Refactor Monitoring Report components and update API paths for user and organization registration
