+++ web/src/Config/apiConfig.ts
@@ -8,6 +8,7 @@ export const API_PATHS = {
   USER_PROFILE: 'User/profile',
   USER_DETAILS: 'user/query',
   ADD_USER: 'user/add',
+  REGISTER_USER: 'user/register',
   UPDATE_USER: 'user/update',
   DELETE_USER: (userId: string) => `user/delete?userId=${userId}`,
   RESET_PASSWORD_USER: 'user/resetPassword',
@@ -45,6 +46,7 @@ export const API_PATHS = {
   REGIONS: 'organisation/regions',
   ORGANIZATION_BY_TYPE: 'organisation/byType',
   ORGANIZATION_DETAILS: 'organisation/query',
+  UPDATE_ORGANIZATION: 'organisation/update',
   ORGANIZATION_NAMES: 'organisation/queryNames',
   DOWNLOAD_ORGANIZATION_DATA: 'organisation/download',
   ORGANIZATION_PROFILE_DETAILS: (companyId: string) => `organisation/profile?id=${companyId}`,
