+++ web/src/locales/i18n/monitoringReport/en.json
@@ -1,5 +1,5 @@
 {
-  "title01": "Project Details",
+  "title01": "Basic information",
   "title02": "Description of Project Activity",
   "title03": "Implementation Status",
   "title04": "Safeguards",
@@ -7,15 +7,26 @@
   "title06": "Quantification of GHG Emission Reductions and Removals",
   "title07": "Annexures",
   "monitoringReport": "Monitoring Report",
-  "pd_projectTitle": "Project Title",
   "pd_reportID": "Report ID",
   "pd_pages": "Pages",
-  "pd_projectProponents": "Project Proponents",
   "pd_physicalAddress": "Physical Address",
   "pd_email": "E-mail",
   "pd_version": "Version",
   "pd_reportTitle": "Report Title",
-  "pd_dateOfIssue": "Date of Issue",
+  "bi_applicablePDDVersionNo": "Version number of the PDD applicable to this monitoring report",
+  "bi_projectTitle": "Title of the Project Activity",
+  "bi_completionDate": "Date of Issue",
+  "bi_duration": "Duration of this monitoring period",
+  "bi_projectParticipants": "Project participants",
+  "bi_appliedMethodologies": "Applied methodologies and standardized baselines",
+  "bi_achievedGHGReductions": "GHG reductions achieved during the monitoring period",
+  "bi_unfccRefNo": "UNFCC reference number of the project activity",
+  "bi_versionNoOfMR": "Version number of this monitoring report",
+  "bi_monitoringPeriodNo": "Monitoring period number",
+  "bi_monitoringNoForMonitoringPeriod": "Monitoring report number for this monitoring period",
+  "bi_hostParty": "Host Party",
+  "bi_sectoralScope": "Sectoral scopes",
+  "bi_projectedGHGReductions": "Projected GHG Reductions",
   "pd_preparedBy": "Prepared By",
   "pd_telephone": "Telephone",
   "pd_website": "Website",
@@ -176,5 +187,4 @@
   "reject": "Reject",
   "approve": "Approve",
   "next": "Next"
-
 }
